requirements.txt
pandas
torch
Pillow
pyautogui
pynput
tqdm
sklearn
matplotlib
